+++ source/SmartSignalFilter/SmartSignalFilter.h
@@ -4,7 +4,7 @@
 class SmartSignalFilter : public SignalFilter {
 	public:
 		SmartSignalFilter(int threadNum, AlgorithmType algType, int maskSize);
-		void apply(std::vector<int>& signal) override;
+		void apply(std::vector<uint8_t>& signal) override;
 	private:
-		void filter(std::vector<int>& newPartOfSignal, int firstIndex, int lastIndex);
+		void filter(std::vector<uint8_t>& newPartOfSignal, int firstIndex, int lastIndex);
 };
